Update KA-15631.md
Hi! I'm Jim Menn, Senior Technical Writer at Adobe, on Director Stephanie Marques' team, working on a KCS project with Roxann McGlumphy/Emily Geary to edit many of these articles, and I made this PR because the Dynamics editor would not allow the formatting that Markdown so easily provides for this article.